Report meminfo stats
When putting the file page cache under pressure, the "MemFree",
"Active(file)" and "Inactive(file)" stats are certainly useful.

For the sake of correlation, make the resident refresher include those in
its regular stat reports.

If the resident refresher is not run, spawn a thread emitting these
every half a second.

Signed-off-by: Nicolai Stange <nstange@suse.de>